--Move SimulatorConfiguration to separate file.
