+++ src/esp/sim/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_library(
   sim STATIC
-  Simulator.cpp Simulator.h
+  Simulator.cpp Simulator.h SimulatorConfiguration.cpp SimulatorConfiguration.h
 )
 
 target_link_libraries(
